[track] Optimize database access in prev_segment.
The prev_segment function was called by a recursive function. This
led to a great amount of redundant reads from mnesia to retrieve the
total number of segments. By moving the get_setting/1 call upper in
the call stack, the number of needed transactions has been reduced
to just 1 for each preelaborate/1 invocation.