[bootstrap_server,node_manager] Rewrite apps' start and shutdown mechanisms.
Because of a limitation in dist_ac (all node names must be known at startup),
we had to redesign a fundamental part of the bootstrap process, in order to
circumvent dist_ac altogether.
This way the shutdown of applications becomes slightly less trivial on the
bootstrap_server's side, and we also lose distributed applications (in the
OTP sense). Luckily this isn't an issue since we're not going to implement
fault tolerance anyway.